backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
bidict==0.23.1
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.8
fastapi-socketio==0.0.10
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
MarkupSafe==3.0.2
motor==3.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
openai==1.65.4
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.4
safetensors==0.5.2
setuptools==3.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tokenizers==0.15.2
torch==2.6.0
tqdm==4.67.1
transformers==4.39.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.29.0
wsproto==1.2.0
PyPDF2==3.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pydantic-settings==2.4.0
python-pptx==0.6.23
Pillow==10.1.0